add simple serializer for layer results and use it in layer view GeoJSONResultLayerData #38
as django-geojson's GeoJSONLayerView cannot be used here. Explanation: at serialization time the queryset must be a django QuerySet to have the 'id' included in the features (feature id needed for popup url). If a list of Datasets is provided (which results from adding an additional (virtual) column/property), it's not. But there's no way of having the id included in the GeoJSON without adjusting django-geojson. so